% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centimetros_a_pulgadas.R
\name{centimetros_a_pulgadas}
\alias{centimetros_a_pulgadas}
\title{Convierte pulgadas a centímetros}
\usage{
centimetros_a_pulgadas(pulgadas)
}
\arguments{
\item{pulgadas}{Valor numérico en pulgadas.}
}
\value{
Valor numérico en centímetros.
}
\description{
Convierte pulgadas a centímetros
}
\examples{
pulgadas_a_centimetros(10)
}
