% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_estaciones.R
\name{leer_estaciones}
\alias{leer_estaciones}
\title{Descargar y leer datos de estaciones meteorológicas}
\usage{
leer_estaciones(ruta_datos = here::here("datos"))
}
\arguments{
\item{ruta_datos}{Directorio donde se guardarán/leerán los archivos
.csv. Por defecto se usa \code{here::here("datos")}.}
}
\value{
Una lista nombrada de data frames, uno por archivo descargado.
}
\description{
Descarga archivos .csv con metadatos y datos de estaciones
meteorológicas desde un repositorio en línea, los guarda en un
directorio y los lee en una lista de data frames.
}
