% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadatos.R
\docType{data}
\name{metadatos}
\alias{metadatos}
\title{Metadatos de estaciones meteorológicas}
\format{
Un data frame con X filas y Y columnas.
}
\source{
Elaboración propia
}
\usage{
metadatos
}
\description{
Data frame con información descriptiva de las estaciones usadas en el paquete.
}
\examples{
data(metadatos)
head(metadatos)
}
\keyword{datasets}
