% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0046.R
\docType{data}
\name{NH0046}
\alias{NH0046}
\title{Datos de la estación NH0046}
\format{
Un data frame con varias columnas de mediciones climáticas.
}
\source{
Elaboración propia
}
\usage{
NH0046
}
\description{
Datos climáticos registrados en la estación NH0046.
}
\examples{
data(NH0046)
head(NH0046)
}
\keyword{datasets}
