% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0098.R
\docType{data}
\name{NH0098}
\alias{NH0098}
\title{Datos de la estación NH0098}
\format{
Un data frame con varias columnas de mediciones climáticas.
}
\source{
Elaboración propia
}
\usage{
NH0098
}
\description{
Datos climáticos registrados en la estación NH0098.
}
\examples{
data(NH0098)
head(NH0098)
}
\keyword{datasets}
