% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0437.R
\docType{data}
\name{NH0437}
\alias{NH0437}
\title{Datos de la estación NH0437}
\format{
Un data frame con varias columnas de mediciones climáticas.
}
\source{
Elaboración propia
}
\usage{
NH0437
}
\description{
Datos climáticos registrados en la estación NH0437.
}
\examples{
data(NH0437)
head(NH0437)
}
\keyword{datasets}
