% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0472.R
\docType{data}
\name{NH0472}
\alias{NH0472}
\title{Datos de la estación NH0472}
\format{
Un data frame con varias columnas de mediciones climáticas.
}
\source{
Elaboración propia
}
\usage{
NH0472
}
\description{
Datos climáticos registrados en la estación NH0472.
}
\examples{
data(NH0472)
head(NH0472)
}
\keyword{datasets}
