% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0910.R
\docType{data}
\name{NH0910}
\alias{NH0910}
\title{Datos de la estación NH0910}
\format{
Un data frame con varias columnas de mediciones climáticas.
}
\source{
Elaboración propia
}
\usage{
NH0910
}
\description{
Datos climáticos registrados en la estación NH0910.
}
\examples{
data(NH0910)
head(NH0910)
}
\keyword{datasets}
